[Event] order buy command 추가
