+++ backend/store_api/src/order/application/order.buy.handler.ts
@@ -2,24 +2,84 @@ import { CommandHandler, EventBus, ICommandHandler } from '@nestjs/cqrs';
 import { OrderBuyCommand } from '../domain/order.buy.command';
 import { Inject } from '@nestjs/common';
 import { IOrderRepository } from '../domain/order.repository';
-import { IOrderProductRepository } from '../domain/orderProduct.repository';
 import { IProductRepository } from '../../product/domain/product.repository';
 import { IOrderPaymentServer } from '../domain/order.payment.server';
 import { CartBuyEvent } from '../../cart/domain/cart.buy.event';
+import { listToMap } from '@libs/utils';
+import { CoPangException, EXCEPTION_STATUS } from '../../common/domain/exception';
+import { createOrderCode, createOrderName, mergeOrderProduct, sumTotalCost } from '../domain/order';
+import { Product } from '../../product/domain/product';
+import { OrderCreateOut, OrderPaymentCreateBuyOut, OrderProductCreateBuyOut } from '../domain/port/order.out';
 
 @CommandHandler(OrderBuyCommand)
 export class OrderBuyHandler implements ICommandHandler<OrderBuyCommand> {
   constructor(
     @Inject('IOrderRepository') private orderRepository: IOrderRepository,
-    @Inject('IOrderProductRepository') private orderProductRepository: IOrderProductRepository,
     @Inject('IProductRepository') private productRepository: IProductRepository,
     @Inject('IOrderPaymentServer') private orderPaymentServer: IOrderPaymentServer,
     private eventBus: EventBus,
   ) {}
 
   async execute(command: OrderBuyCommand) {
-    // TODO Service Logic
-    console.log(command);
-    this.eventBus.publish(new CartBuyEvent(1, []));
+    const buyIn = command.orderBuyIn;
+    const buyerId = buyIn.buyerId;
+
+    const productIdMap = listToMap(buyIn.buyProduct, (product) => product.productId);
+    const productIdArray = Array.from(productIdMap.keys());
+
+    const productArray = await this.productRepository.findAllById(productIdArray);
+
+    if (productArray.length !== productIdArray.length) {
+      throw new CoPangException(EXCEPTION_STATUS.PRODUCT_NOT_EXIST);
+    }
+
+    const mergeProductMap = mergeOrderProduct(productArray, productIdMap);
+    const mergeProductArray = Array.from(mergeProductMap.values());
+
+    for (const product of mergeProductArray) {
+      if (!product.isSale) {
+        throw new CoPangException(EXCEPTION_STATUS.PRODUCT_NOT_AVAILABLE_BUY);
+      }
+      if (Product.isDeleted(product)) {
+        throw new CoPangException(EXCEPTION_STATUS.PRODUCT_NOT_EXIST);
+      }
+      if (Product.isOverQuantity(product, product.buyQuantity)) {
+        throw new CoPangException(EXCEPTION_STATUS.PRODUCT_NOT_AVAILABLE_BUY);
+      }
+    }
+
+    const productNameArray = productArray.map((product) => product.name);
+    const totalCost = mergeProductArray.reduce(sumTotalCost, 0);
+    const code = createOrderCode();
+    const name = createOrderName(productNameArray);
+    const now = new Date();
+    const address = buyIn.address;
+
+    const paymentResult = await this.orderPaymentServer.request({ ...buyIn.card });
+
+    const order: OrderCreateOut = {
+      code,
+      name,
+      totalCost,
+    };
+    const payment: OrderPaymentCreateBuyOut = {
+      ...buyIn.card,
+
+      orderCode: code,
+      orderName: name,
+      paymentKey: paymentResult.paymentKey,
+      requestAt: now,
+      totalAmount: totalCost,
+    };
+    const orderProduct: OrderProductCreateBuyOut[] = mergeProductArray.map((product) => {
+      return {
+        ...product,
+        productId: product.id,
+        address,
+      };
+    });
+
+    await this.orderRepository.buy({ buyerId: buyIn.buyerId, order: order, payment: payment, buyProduct: orderProduct });
+    await this.eventBus.publish(new CartBuyEvent(buyerId, productIdArray));
   }
 }
